1.优化制作图片界面删除单张图片时，同步删除sd卡对应的图片 2.sd卡图片被删除时，修改图片同步删除sqlite对应的记录

+++ app/src/main/java/com/yzi/doutu/utils/ImageUtils.java
@@ -535,7 +535,7 @@ public class ImageUtils {
         }
     }
 
-    public static String saveBitmapToFiles(Bitmap tempBitmap,final DataBean dataBean,int height,int width) {
+    public static String saveBitmapToFiles(Bitmap tempBitmap,final DataBean dataBean) {
         if (TextUtils.isEmpty(getFilesPath(FILE_DIY_PATH))) {
             return null;
         }
@@ -551,11 +551,14 @@ public class ImageUtils {
             out.flush();
             out.close();
 
-            //保证缓存图片的唯一性删掉旧图,替换新的文件名
-            if(!TextUtils.isEmpty(dataBean.getFileName())){
-                SimpleFileUtils.deleteFile(new File(FILE_DIY_PATH+dataBean.getFileName()),null);
+            if(dataBean!=null){
+                //保证缓存图片的唯一性删掉旧图,替换新的文件名
+                if(!TextUtils.isEmpty(dataBean.getFileName())){
+                    SimpleFileUtils.deleteFile(new File(FILE_DIY_PATH+dataBean.getFileName()),null);
+                }
+                dataBean.setFileName(fileName);
             }
-            dataBean.setFileName(fileName);
+
 
         } catch (Exception e) {
             e.printStackTrace();
